Minor refactors; list of updates from previous commit in description
Using char array instead of byte array because chars are unsigned.
Rewrote the array reordering function to be cleaner
Instead of making and saving a .jpg, rectangles are just drawn and
filled for each pixel.
Math is done to ignore the first two bits of every pixel byte.
HTML file is implemented that embeds the applet.